feat: add other blocks
