+++ .changeset/warm-rabbits-run.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": minor
+---
+
+Add blocks into new Market page
