+++ apps/evm/src/pages/Market/Market/MarketHistory/index.tsx
@@ -11,7 +11,7 @@ import {
 } from 'utilities';
 
 import TEST_IDS from '../../testIds';
-import { MarketCard, type MarketCardProps } from '../../Market/MarketCard';
+import { MarketCard, type MarketCardProps } from '../MarketCard';
 import useGetChartData from './useGetChartData';
 
 interface MarketHistoryProps {
